// Dashboard/src/modules/common/models/location.ts
/*
 * Copyright 2022 IBM Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {Executor} from './executor';

export interface AdminLevel {
  levelId?: number;
  levelName?: string;
}

export interface Location {
  id?: string;
  names?: string;
  adminType?: string;
  adminLevel?: string; //
  adminLevelList?: AdminLevel[]; //
  locatedIn?: string;
  country?: string;
  active?: boolean;
  verified?: boolean;
  metadataDetailsList?: MetadataDetails[];
  locationDataId?: string;
}

export interface MetadataDetails {
  contentType?: string;
  createdAt?: number;
  dataRepositoryConfiguration?: DataRepositoryConfiguration;
  description?: string;
  executorId?: string;
  id?: string;
  key?: string;
  locationId?: string;
  maxDays?: string;
  name?: string;
  source?: string;
  startDate?: string;
  testData?: string;
  type?: string;
  updatedAt?: number;
}

export interface DataRepositoryConfiguration {
  createdAt?: number;
  credentials?: string;
  description?: string;
  hostEndPoint?: string;
  id?: string;
  name?: string;
  updatedAt?: number;
  sourceUrl?: string;
  fileBytesStream?: FormData;
}

export interface LocationData {
  id?: string;
  metadataDetailsList?: MetadataDetails[];
  location?: Location;
  executorList?: Executor[];
}